syncing prior to thesis impl

+++ include/net/ip_puzzle.h
@@ -3,7 +3,7 @@
 
 struct inet_solution;
 
-struct inet_solution *solve_ip_puzzle(int difficulty, int pnum,
+struct inet_solution *solve_ip_puzzle(int difficulty, u8 pnum,
 				      __be32 ts, u8 *s_nonce, u8 *c_nonce);
 
 #endif
